\name{edgelist2matrix}
\alias{edgelist2matrix}
\title{Make a matrix from an edgelist on file.}
\usage{
  edgelist2matrix(path)
}
\arguments{
  \item{path}{Path to a edgelist file.}
}
\description{
  Make a matrix from an edgelist on file.
}
\examples{
edgelist2matrix("/Users/scottmac2/testtest/sp30_bal_ou_comp_1.web")
}

