\name{pbtree}
\alias{pbtree}
\title{Written by Liam J. Revell 2011
From the phytools package
citation: Revell, L. J. (2012) phytools: An R package for phylogenetic comparative biology (and other things) Methods Ecol. Evol. 3, 217-223 doi:10.1111/j.2041-210X.2011.00169.x}
\usage{
  pbtree(b = 1, n = NULL, t = NULL, scale = NULL, nsim = 1,
    ape = TRUE)
}
\description{
  Written by Liam J. Revell 2011 From the phytools package
  citation: Revell, L. J. (2012) phytools: An R package for
  phylogenetic comparative biology (and other things)
  Methods Ecol. Evol. 3, 217-223
  doi:10.1111/j.2041-210X.2011.00169.x
}
\keyword{internal}

