\name{sim_rand_nets}
\alias{sim_rand_nets}
\title{Measure aspects of traits on trees.}
\usage{
  sim_rand_nets(listoftrees)
}
\arguments{
  \item{listoftrees}{}
}
\description{
  Measure aspects of traits on trees.
}
\examples{
\dontrun{
trees <- list(rmtree(N=50, n=10), rmtree(N=50, n=20))
sim_rand_nets(trees)
}
}

